Refactoring `ForwardAnalysis`
Break it into multiple files, to make it tidier.
